requirements.txt
flask_jwt_extended
flask_sqlalchemy
bcrypt
psycopg2
Flask
python-dotenv